Create code
In this function, you need to pass three parameters: input1, input2, and operator. These represent the values and operator you want to use in your engineering calculation. The function will then use fetch() to make a request to the engineering calculator API with the provided inputs. The response is then parsed as JSON and returned.

Note that this function uses the fetch() method, which is a newer feature in JavaScript that may not be supported in all browsers. If you need to support older browsers, you may need to use a polyfill or a different approach to making HTTP requests.